darkColor on editGroup+ repair little problem

+++ app/src/main/java/com/example/knocker/controller/activity/EditContactActivity.kt
@@ -599,12 +599,18 @@ class EditContactActivity : AppCompatActivity() {
     }
 
     private fun deleteContact() {
-        edit_contact_ContactsDatabase!!.contactsDao().deleteContactById(edit_contact_id!!)
-        val mainIntent = Intent(this@EditContactActivity, MainActivity::class.java)
-        mainIntent.putExtra("isDelete", true)
-        startActivity(mainIntent)
-        finish()
+        MaterialAlertDialogBuilder(this, R.style.AlertDialog)
+                .setTitle(getString(R.string.edit_contact_delete_contact))
+                .setMessage(getString(R.string.edit_contact_delete_contact_message))
+                .setPositiveButton("Remove") { _, _ ->
+                    edit_contact_ContactsDatabase!!.contactsDao().deleteContactById(edit_contact_id!!)
+                    val mainIntent = Intent(this@EditContactActivity, MainActivity::class.java)
+                    mainIntent.putExtra("isDelete", true)
+                    startActivity(mainIntent)
+                    finish()
+                }.show()
     }
+
 
     private fun hideKeyboard() {
         // Check if no view has focus:
